+++ test/src/unit-filter-pipeline.cc
@@ -69,6 +69,10 @@ class Add1InPlace : public Filter {
       : Filter(FilterType::FILTER_NONE) {
   }
 
+  void dump(FILE* out) const override {
+    (void)out;
+  }
+
   Status run_forward(
       FilterBuffer* input_metadata,
       FilterBuffer* input,
@@ -127,6 +131,10 @@ class Add1OutOfPlace : public Filter {
   // Just use a dummy filter type
   Add1OutOfPlace()
       : Filter(FilterType::FILTER_NONE) {
+  }
+
+  void dump(FILE* out) const override {
+    (void)out;
   }
 
   Status run_forward(
@@ -211,6 +219,10 @@ class AddNInPlace : public Filter {
     increment_ = 1;
   }
 
+  void dump(FILE* out) const override {
+    (void)out;
+  }
+
   Status run_forward(
       FilterBuffer* input_metadata,
       FilterBuffer* input,
@@ -282,6 +294,11 @@ class PseudoChecksumFilter : public Filter {
   PseudoChecksumFilter()
       : Filter(FilterType::FILTER_NONE) {
   }
+
+  void dump(FILE* out) const override {
+    (void)out;
+  }
+
   Status run_forward(
       FilterBuffer* input_metadata,
       FilterBuffer* input,
@@ -360,6 +377,10 @@ class Add1IncludingMetadataFilter : public Filter {
   // Just use a dummy filter type
   Add1IncludingMetadataFilter()
       : Filter(FilterType::FILTER_NONE) {
+  }
+
+  void dump(FILE* out) const override {
+    (void)out;
   }
 
   Status run_forward(