+++ tiledb/sm/c_api/tiledb.cc
@@ -1536,7 +1536,7 @@ int32_t tiledb_attribute_get_filter_list(
 
   // Create a new FilterPipeline object
   (*filter_list)->pipeline_ =
-      new (std::nothrow) tiledb::sm::FilterPipeline(*attr->attr_->filters());
+      new (std::nothrow) tiledb::sm::FilterPipeline(attr->attr_->filters());
   if ((*filter_list)->pipeline_ == nullptr) {
     delete *filter_list;
     auto st = tiledb::sm::Status::Error(
@@ -1728,6 +1728,70 @@ void tiledb_dimension_free(tiledb_dimension_t** dim) {
     delete *dim;
     *dim = nullptr;
   }
+}
+
+int32_t tiledb_dimension_set_filter_list(
+    tiledb_ctx_t* ctx,
+    tiledb_dimension_t* dim,
+    tiledb_filter_list_t* filter_list) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, dim) == TILEDB_ERR ||
+      sanity_check(ctx, filter_list) == TILEDB_ERR)
+    return TILEDB_ERR;
+
+  if (SAVE_ERROR_CATCH(
+          ctx, dim->dim_->set_filter_pipeline(filter_list->pipeline_)))
+    return TILEDB_ERR;
+
+  return TILEDB_OK;
+}
+
+int32_t tiledb_dimension_set_cell_val_num(
+    tiledb_ctx_t* ctx, tiledb_dimension_t* dim, uint32_t cell_val_num) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, dim) == TILEDB_ERR)
+    return TILEDB_ERR;
+  if (SAVE_ERROR_CATCH(ctx, dim->dim_->set_cell_val_num(cell_val_num)))
+    return TILEDB_ERR;
+  return TILEDB_OK;
+}
+
+int32_t tiledb_dimension_get_filter_list(
+    tiledb_ctx_t* ctx,
+    tiledb_dimension_t* dim,
+    tiledb_filter_list_t** filter_list) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, dim) == TILEDB_ERR)
+    return TILEDB_ERR;
+
+  // Create a filter list struct
+  *filter_list = new (std::nothrow) tiledb_filter_list_t;
+  if (*filter_list == nullptr) {
+    auto st = tiledb::sm::Status::Error(
+        "Failed to allocate TileDB filter list object");
+    LOG_STATUS(st);
+    save_error(ctx, st);
+    return TILEDB_OOM;
+  }
+
+  // Create a new FilterPipeline object
+  (*filter_list)->pipeline_ =
+      new (std::nothrow) tiledb::sm::FilterPipeline(dim->dim_->filters());
+  if ((*filter_list)->pipeline_ == nullptr) {
+    delete *filter_list;
+    auto st = tiledb::sm::Status::Error(
+        "Failed to allocate TileDB filter list object");
+    LOG_STATUS(st);
+    save_error(ctx, st);
+    return TILEDB_OOM;
+  }
+
+  return TILEDB_OK;
+}
+
+int32_t tiledb_dimension_get_cell_val_num(
+    tiledb_ctx_t* ctx, const tiledb_dimension_t* dim, uint32_t* cell_val_num) {
+  if (sanity_check(ctx) == TILEDB_ERR || sanity_check(ctx, dim) == TILEDB_ERR)
+    return TILEDB_ERR;
+  *cell_val_num = dim->dim_->cell_val_num();
+  return TILEDB_OK;
 }
 
 int32_t tiledb_dimension_get_name(
@@ -2186,8 +2250,8 @@ int32_t tiledb_array_schema_get_coords_filter_list(
   }
 
   // Create a new FilterPipeline object
-  (*filter_list)->pipeline_ = new (std::nothrow) tiledb::sm::FilterPipeline(
-      *array_schema->array_schema_->coords_filters());
+  (*filter_list)->pipeline_ = new (std::nothrow)
+      tiledb::sm::FilterPipeline(array_schema->array_schema_->coords_filters());
   if ((*filter_list)->pipeline_ == nullptr) {
     delete *filter_list;
     auto st = tiledb::sm::Status::Error(
@@ -2220,7 +2284,7 @@ int32_t tiledb_array_schema_get_offsets_filter_list(
 
   // Create a new FilterPipeline object
   (*filter_list)->pipeline_ = new (std::nothrow) tiledb::sm::FilterPipeline(
-      *array_schema->array_schema_->cell_var_offsets_filters());
+      array_schema->array_schema_->cell_var_offsets_filters());
   if ((*filter_list)->pipeline_ == nullptr) {
     delete *filter_list;
     auto st = tiledb::sm::Status::Error(