+++ tiledb/sm/filter/checksum_md5_filter.h
@@ -75,6 +75,9 @@ class ChecksumMD5Filter : public Filter {
    */
   ChecksumMD5Filter();
 
+  /** Dumps the filter details in ASCII format in the selected output. */
+  void dump(FILE* out) const override;
+
   /**
    * Encrypt the bytes of the input data into the output data buffer.
    */