+++ tiledb/sm/filter/filter_pipeline.cc
@@ -488,12 +488,26 @@ Status FilterPipeline::deserialize(ConstBuffer* buff) {
   return Status::Ok();
 }
 
+void FilterPipeline::dump(FILE* out) const {
+  if (out == nullptr)
+    out = stdout;
+
+  for (const auto& filter : filters_) {
+    fprintf(out, "\n  > ");
+    filter->dump(out);
+  }
+}
+
 void FilterPipeline::set_max_chunk_size(uint32_t max_chunk_size) {
   max_chunk_size_ = max_chunk_size;
 }
 
 unsigned FilterPipeline::size() const {
   return static_cast<unsigned>(filters_.size());
+}
+
+bool FilterPipeline::empty() const {
+  return filters_.empty();
 }
 
 void FilterPipeline::swap(FilterPipeline& other) {